requirements.txt
Django==1.6
South==0.8.2
argparse==1.2.1
django-appconf==0.6
django-braces==1.2.2
django-compressor==1.3
django-crispy-forms==1.4.0
django-datatables-view==1.6
django-debug-toolbar==0.11.0
django-floppyforms==1.1
factory-boy==2.2.1
gunicorn==18.0
newrelic==2.6.0.5
psycopg2==2.5.1
pytz==2013.7
six==1.4.1
sqlparse==0.1.10
wsgiref==0.1.2
